1. Add percentageOnWork to manually edit receipt page 2. Calculate and show GST total and work related totals in Report page 3. Support delete action for archived receipts

+++ lib/data_model/receipt.dart
@@ -26,7 +26,7 @@ class ReceiptListItem {
   int decodeStatus;
   double altTotalAmount;
   String altCurrencyCode;
-  int percentageOnWork = 100;
+  double percentageOnWork = 100;
 
   ReceiptListItem();
